Remove shodan.io API key (which was present in ReconDog)
github.com/UltimateHackers/ReconDog, the project that this is based on,
has had no GitHub issues, no documentation, and no code comments
suggesting users to _remove_ the key, and to add their own. For this
reason I kept the API key in the project; I have removed it. For a
new API key, visit shodan.io, register, and grab yourself an API key.